// packages/lego-webapp/redux/models/User.ts
export interface PhotoConsent {
  year: number;
  semester: Semester;
  domain: PhotoConsentDomain;
  isConsenting: boolean | null;
  updatedAt: Dateish;
}

export interface Achievement {
  id: number;
  identifier: string;
  updatedAt: Dateish;
  level: number;
  percentage: number;
}

export interface AchievementData {
  name: string;
  description: string;
  rarity: number;
  hidden: boolean;
}

export const Gender = {
  male: 'Mann',
  female: 'Kvinne',
  other: 'Annet',
} as const;

export type UserPermissionGroup = Pick<PublicGroup, 'id' | 'name'>;

interface User {
  id: EntityId;
  username: string;
  firstName: string;
  lastName: string;
  fullName: string;
  gender: keyof typeof Gender;
  email: string;
  emailAddress: string;
  emailListsEnabled: boolean;
  internalEmailAddress?: string;
  phoneNumber?: string;
  profilePicture: string;
  profilePicturePlaceholder: string;
  allergies: string;
  isActive: boolean;
  isStudent: boolean;
  abakusEmailLists: PublicEmailList[];
  penalties: EntityId[];
  icalToken: string;
  abakusGroups: EntityId[];
  isAbakusMember: boolean;
  isAbakomMember: boolean;
  pastMemberships: PastMembership[];
  selectedTheme: string;
  permissionsPerGroup: {
    abakusGroup: UserPermissionGroup;
    permissions: string[];
    parentPermissions: {
      abakusGroup: UserPermissionGroup;
      permissions: string[];
    }[];
  }[];
  photoConsents?: PhotoConsent[];
  memberships: Membership[];
  githubUsername?: string;
  linkedinId?: string;
  actionGrant?: ActionGrant;
  achievements: Achievement[];
  achievementsScore: number;
}
// ...
export type AutocompleteUser = Pick<
  User,
  'username' | 'fullName' | 'profilePicture' | 'id'
> & {
  contentType: AutocompleteContentType.User;
  text: 'text';
};